Fixed all the problems with Lock, Unlock and the blocker subsystem
