+++ commands/unlock.py
@@ -2,12 +2,12 @@ from utils.safecommand import SafeCommand
 from subsystems.blocker import Blocker
 
 
-class CloseLock(SafeCommand):
+class Unlock(SafeCommand):
     def __init__(self, blocker: Blocker):
         super().__init__()
         self.blocker = blocker
-        self.addRequirements(blocker)
+        self.addRequirements(self.blocker)
 
     def execute(self) -> None:
-        if self.blocker.limit_switch_blocker.get():
-            self.blocker.lock()
+        if not self.blocker.isArmDown():
+            self.blocker.unlock()
